requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
requests>=2.31.0
python-dotenv>=1.0.0
celery[redis]>=5.3.6
redis>=5.0.3
pydantic>=2.0.0
